+++ api/src/apis/index.ts
@@ -1,0 +1,2 @@
+export { VaraApi } from './vara/api';
+export { VaraTestnetApi } from './vara-testnet/api';