+++ api/src/index.ts
@@ -19,3 +19,5 @@ export * from './utils';
 export * from './events';
 export * from './types';
 export * from './metadata';
+export * from './apis';
+export * from './specs';